Move access point interfaces to authclient. Aliases left behind to prevent breaking builds
